Revamped the whole parsing mechanism.
The previous way was too limited. After analysing the XML structure further, a lightweight parser was implemented that keeps track of the structure of the document. This allows searching by tag path, or by tag names with a specific set of attributes.

Every tag that can be found is given a dedicated class.